Quy add

+++ spring-boot/src/main/java/codegym/module4/entities/Ticket.java
@@ -1,8 +1,5 @@
 package codegym.module4.entities;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
 import javax.persistence.*;
 
 
@@ -25,40 +22,38 @@ public class Ticket {
     @Column(name = "booked_ticket_code")
     private String bookedTicketCode;
 
+
+
     @Column(name = "order_status")
     private String orderStatus;
 
+
+
     @Column(name = "price")
-    private double price;
+    private int price;
 
     @ManyToOne
     @JoinColumn(name = "seat_id")
-
     private Seat seat;
 
     @ManyToOne
     @JoinColumn(name = "ticket_price_id")
-
     private TicketPrices ticketPrices;
 
     @ManyToOne
     @JoinColumn(name = "promotion_id")
-
     private Promotion promotion;
 
     @ManyToOne
     @JoinColumn(name = "customer_id")
-
     private Customer customer;
 
     @ManyToOne
     @JoinColumn(name = "employee_id")
-
     private Employee employee;
 
     @ManyToOne
     @JoinColumn(name = "movie_schedule_id")
-
     private MovieSchedules movieSchedules;
 
 
@@ -126,14 +121,14 @@ public class Ticket {
     /**
      * @return the price
      */
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
     /**
      * @param price the price to set
      */
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
